requirements.txt
flask
tensorflow-cpu
numpy
Pillow
gunicorn